refactor: Move `PropertyPath` to one file
- Remove now unneeded headers from `id_types.hpp`
- Forward declare `PropertyPath` where we can
