+++ src/storage/v2/property_store.cpp
@@ -24,6 +24,7 @@
 #include <utility>
 
 #include "storage/v2/id_types.hpp"
+#include "storage/v2/indices/property_path.hpp"
 #include "storage/v2/property_value.hpp"
 #include "storage/v2/temporal.hpp"
 #include "utils/cast.hpp"
